src/packing-algo/requirements.txt
Flask==2.1.1
gunicorn==20.1.0
python-socketio==5.3.0
supabase==2.7.3
python-dotenv
werkzeug==2.1.1
pandas